Added fizzbuzz_math_tricks.lua Instead of just modulous, why not use math tricks to determine FizzBuzz?
For determining Fizz, the number has it's digits summed in a recursive
manner. Once reduced to a single digit, I use a small lookup table to
determine if the recursive sum of digits resulted in a number divisible
by 3.

For determining Buzz, I check a small lookup table with the last digit
in the number. If the last digit is 5 or 0 the number is divisible by 5.

These "math trick" approaches are in methods for for .fizz() and .buzz()